refactor(reader): centralize malloced variables in reader struct
• Refactored `t_reader` struct to include all malloced variables:
  - Added `cached_input`, `linereader`, `commands`, and `varnames` fields.
• Updated `reader_init` to initialize new fields and `reader_free` to handle their cleanup.
• Modified `main.c` to use `t_reader` for memory management, replacing local variables.
• Introduced `free_command_contents` in `parser_free.c` for consistent command cleanup.
• Adjusted `Makefile` to include new source files: `main_extra.c` and `remove_empty_commands.c`.

This change improves memory management by consolidating all dynamically allocated variables into the `t_reader` struct, ensuring proper cleanup and reducing the risk of memory leaks.
