+++ include/shared.h
@@ -6,7 +6,7 @@
 /*   By: lfiorell@student.42nice.fr <lfiorell>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/14 14:24:32 by lfiorell@st       #+#    #+#             */
-/*   Updated: 2025/07/16 17:23:36 by jfranc           ###   ########.fr       */
+/*   Updated: 2025/07/16 19:32:12 by lfiorell@st      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
